compare-openai-agents-to-openai-sdk/openai-agents-sdk/pyproject.toml
[project]
name = "openai-chat"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.0.3",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "python-dotenv>=1.0.0",
    "aiosqlite>=0.20.0",
]
